Refactor Coulomb scattering data
- Add a new data/params class for Wentzel OK&VI data that will be used by both
  the single Coulomb scattering model and the Wentzel VI MSC model
- Import MscThetaLimit from EM parameters instead of the model PolarAngleLimit
