+++ src/celeritas/em/data/WentzelVIMscData.hh
@@ -57,8 +57,6 @@ struct WentzelVIMscData
     units::MevMass electron_mass;
     //! User-assignable options
     WentzelVIMscParameters params;
-    //! Parameters used in both single and multiple scattering
-    CoulombParameters coulomb_params;
     //! Scaled xs data
     Items<XsGridData> xs;  //!< [mat][particle]
 
@@ -70,8 +68,8 @@ struct WentzelVIMscData
     //! Check whether the data is assigned
     explicit CELER_FUNCTION operator bool() const
     {
-        return ids && electron_mass > zero_quantity() && coulomb_params
-               && !xs.empty() && !reals.empty();
+        return ids && electron_mass > zero_quantity() && !xs.empty()
+               && !reals.empty();
     }
 
     //! Assign from another set of data
@@ -82,7 +80,6 @@ struct WentzelVIMscData
         ids = other.ids;
         electron_mass = other.electron_mass;
         params = other.params;
-        coulomb_params = other.coulomb_params;
         xs = other.xs;
         reals = other.reals;
         return *this;
