+++ src/celeritas/em/model/CoulombScatteringModel.cu
@@ -8,6 +8,7 @@
 #include "CoulombScatteringModel.hh"
 
 #include "celeritas/em/executor/CoulombScatteringExecutor.hh"
+#include "celeritas/em/params/WentzelOKVIParams.hh"
 #include "celeritas/global/ActionLauncher.device.hh"
 #include "celeritas/global/CoreParams.hh"
 #include "celeritas/global/CoreState.hh"
@@ -23,11 +24,14 @@ namespace celeritas
 void CoulombScatteringModel::execute(CoreParams const& params,
                                      CoreStateDevice& state) const
 {
+    CELER_EXPECT(params.wentzel());
+
     auto execute = make_action_track_executor(
         params.ptr<MemSpace::native>(),
         state.ptr(),
         this->action_id(),
-        InteractionApplier{CoulombScatteringExecutor{this->device_ref()}});
+        InteractionApplier{CoulombScatteringExecutor{
+            this->device_ref(), params.wentzel()->device_ref()}});
     static ActionLauncher<decltype(execute)> const launch_kernel(*this);
     launch_kernel(state, execute);
 }
