+++ src/celeritas/em/model/CoulombScatteringModel.hh
@@ -9,9 +9,7 @@
 
 #include <memory>
 
-#include "corecel/data/CollectionMirror.hh"
 #include "celeritas/em/data/CoulombScatteringData.hh"
-#include "celeritas/phys/AtomicNumber.hh"
 #include "celeritas/phys/ImportedModelAdapter.hh"
 #include "celeritas/phys/ImportedProcessAdapter.hh"
 #include "celeritas/phys/Model.hh"
@@ -32,46 +30,12 @@ class CoulombScatteringModel final : public Model
     //!@{
     //! \name Type aliases
     using SPConstImported = std::shared_ptr<ImportedProcesses const>;
-    using HostRef = CoulombScatteringHostRef;
-    using DeviceRef = CoulombScatteringDeviceRef;
     //!@}
-
-    //! Wentzel Coulomb scattering model configuration options
-    struct Options
-    {
-        //! Use combined single and multiple scattering
-        bool is_combined{true};
-
-        //! Polar angle limit between single and multiple scattering
-        real_type polar_angle_limit{constants::pi};
-
-        //! Factor for dynamic computation of angular limit between SS and MSC
-        real_type angle_limit_factor{1};
-
-        //! Nuclear form factor model
-        NuclearFormFactorType form_factor_model{
-            NuclearFormFactorType::exponential};
-
-        //! User defined screening factor
-        real_type screening_factor{1};
-
-        //! Whether to use integral method to sample interaction length
-        bool use_integral_xs{true};
-
-        //! Check if the options are valid
-        explicit operator bool() const
-        {
-            return polar_angle_limit >= 0 && polar_angle_limit <= constants::pi
-                   && angle_limit_factor > 0 && screening_factor > 0;
-        }
-    };
 
   public:
     // Construct from model ID and other necessary data
     CoulombScatteringModel(ActionId id,
                            ParticleParams const& particles,
-                           MaterialParams const& materials,
-                           Options const& options,
                            SPConstImported data);
 
     // Particle types and energy ranges that this model applies to
@@ -100,24 +64,13 @@ class CoulombScatteringModel final : public Model
 
     //!@{
     //! Access model data
-    HostRef const& host_ref() const { return data_.host_ref(); }
-    DeviceRef const& device_ref() const { return data_.device_ref(); }
+    CoulombScatteringData const& host_ref() const { return data_; }
+    CoulombScatteringData const& device_ref() const { return data_; }
     //!@}
 
   private:
-    CollectionMirror<CoulombScatteringData> data_;
+    CoulombScatteringData data_;
     ImportedModelAdapter imported_;
-
-    // Construct per-material and per-element data (loads Mott coefficients)
-    void build_data(HostVal<CoulombScatteringData>& host_data,
-                    MaterialParams const& materials);
-
-    // Retrieve matrix of interpolated Mott coefficients
-    static CoulombScatteringElementData::MottCoeffMatrix
-    get_mott_coeff_matrix(AtomicNumber z);
-
-    // Calculate the nuclear form prefactor
-    static real_type calc_nuclear_form_prefactor(IsotopeView const& iso);
 };
 
 //---------------------------------------------------------------------------//
