+++ src/celeritas/em/params/WentzelOKVIParams.cc
@@ -1,0 +1,931 @@
+//----------------------------------*-C++-*----------------------------------//
+// Copyright 2024 UT-Battelle, LLC, and other Celeritas developers.
+// See the top-level COPYRIGHT file for details.
+// SPDX-License-Identifier: (Apache-2.0 OR MIT)
+//---------------------------------------------------------------------------//
+//! \file celeritas/em/params/WentzelOKVIParams.cc
+//---------------------------------------------------------------------------//
+#include "WentzelOKVIParams.hh"
+
+#include <algorithm>
+#include <utility>
+
+#include "corecel/io/Logger.hh"
+#include "corecel/sys/ScopedMem.hh"
+#include "celeritas/em/params/detail/MscParamsHelper.hh"
+#include "celeritas/grid/XsCalculator.hh"
+#include "celeritas/io/ImportData.hh"
+#include "celeritas/mat/MaterialParams.hh"
+
+namespace celeritas
+{
+//---------------------------------------------------------------------------//
+/*!
+ * Construct if Wentzel VI or Coulomb is present, else return nullptr
+ */
+std::shared_ptr<WentzelOKVIParams>
+WentzelOKVIParams::from_import(ImportData const& data,
+                               SPConstMaterials materials)
+{
+    CELER_EXPECT(materials);
+
+    bool wentzel = has_msc_model(data, ImportModelClass::wentzel_vi_uni);
+    bool coulomb = has_model(data, ImportModelClass::e_coulomb_scattering);
+    if (!(wentzel || coulomb))
+    {
+        // No Wentzel VI MSC or Coulomb scattering present
+        return nullptr;
+    }
+
+    // Use combined single and multiple Coulomb scattering if both the single
+    // scattering and the Wentzel VI models are present
+    Options opts;
+    opts.is_combined = wentzel && coulomb;
+    opts.polar_angle_limit = [&] {
+        if (!coulomb)
+        {
+            // Set the maximum scattering angle for Wentzel VI MSC
+            return constants::pi;
+        }
+        if (!wentzel)
+        {
+            // Set the minimum scattering angle for Coulomb single scattering
+            return real_type(0);
+        }
+        // Polar angle limit between single and multiple scattering
+        return data.em_params.msc_theta_limit;
+    }();
+    opts.screening_factor = data.em_params.screening_factor;
+    opts.angle_limit_factor = data.em_params.angle_limit_factor;
+
+    return std::make_shared<WentzelOKVIParams>(materials, opts);
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Construct from cross section data and material properties.
+ */
+WentzelOKVIParams::WentzelOKVIParams(SPConstMaterials materials,
+                                     Options options)
+{
+    CELER_EXPECT(materials);
+
+    ScopedMem record_mem("WentzelOKVIParams.construct");
+
+    HostVal<WentzelOKVIData> host_data;
+
+    host_data.params.is_combined = options.is_combined;
+    host_data.params.costheta_limit = std::cos(options.polar_angle_limit);
+    host_data.params.a_sq_factor
+        = real_type(0.5)
+          * ipow<2>(options.angle_limit_factor * constants::hbar_planck
+                    * constants::c_light * units::femtometer);
+    host_data.params.screening_factor = options.screening_factor;
+    // TODO: import from Geant4
+    host_data.params.form_factor_type = options.form_factor;
+
+    // Load Mott coefficients
+    build_data(host_data, *materials);
+    CELER_ASSERT(host_data);
+
+    // Move to mirrored data, copying to device
+    data_ = CollectionMirror<WentzelOKVIData>{std::move(host_data)};
+    CELER_ENSURE(data_);
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Load Mott coefficients and construct per-element data.
+ */
+void WentzelOKVIParams::build_data(HostVal<WentzelOKVIData>& host_data,
+                                   MaterialParams const& materials)
+{
+    // Build element data
+    size_type const num_elements = materials.num_elements();
+    auto elem_data = make_builder(&host_data.elem_data);
+    elem_data.reserve(num_elements);
+
+    for (auto el_id : range(ElementId{num_elements}))
+    {
+        // Load Mott coefficients
+        MottElementData z_data;
+        z_data.mott_coeff
+            = get_mott_coeff_matrix(materials.get(el_id).atomic_number());
+        elem_data.push_back(z_data);
+    }
+
+    auto prefactors = make_builder(&host_data.nuclear_form_prefactor);
+    prefactors.reserve(materials.num_isotopes());
+    for (auto iso_id : range(IsotopeId{materials.num_isotopes()}))
+    {
+        prefactors.push_back(
+            this->calc_nuclear_form_prefactor(materials.get(iso_id)));
+    }
+    CELER_ENSURE(host_data.nuclear_form_prefactor.size()
+                 == materials.num_isotopes());
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Calculate the constant prefactors of the squared momentum transfer.
+ *
+ * This factor is used in the exponential and Gaussian nuclear form models: see
+ * Eqs. 2.262--2.264 of [LR15].
+ *
+ * Specifically, it calculates \f$ (r_n/\bar h)^2 / 12 \f$. A special case is
+ * inherited from Geant for hydrogen targets.
+ */
+real_type WentzelOKVIParams::calc_nuclear_form_prefactor(IsotopeView const& iso)
+{
+    if (iso.atomic_number().get() == 1)
+    {
+        // TODO: Geant4 hardcodes a different prefactor for hydrogen
+        return real_type{1.5485e-6};
+    }
+
+    // The ratio has units of (MeV/c)^-2, so it's easier to convert the
+    // inverse which has units of MomentumSq, then invert afterwards
+    constexpr real_type ratio
+        = 1
+          / native_value_to<units::MevMomentumSq>(
+                12
+                * ipow<2>(constants::hbar_planck
+                          / (real_type(1.27) * units::femtometer)))
+                .value();
+    return ratio
+           * fastpow(real_type(iso.atomic_mass_number().get()),
+                     2 * real_type(0.27));
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Get interpolated Mott coefficients for an element.
+ *
+ * These coefficients are used by the Lijian, Quing, Zhengming
+ * expression [PRM 8.48] for atomic numbers 1 <= Z <= 92.
+ * This data was taken from Geant4's G4MottData.hh file.
+ *
+ * For higher Z values, the PRM cites a numerical solution by Boschini
+ * et al (2013), but does not implement it.
+ */
+MottElementData::MottCoeffMatrix
+WentzelOKVIParams::get_mott_coeff_matrix(AtomicNumber z)
+{
+    CELER_EXPECT(z);
+
+    using CoeffMat = MottElementData::MottCoeffMatrix;
+    // clang-format off
+    static CoeffMat const mott_coeffs[] = {
+        //H.......................................
+        CoeffMat{{
+        {1.e0,2.67363e-8,7.1153e-8,-9.7703e-8,-6.69132e-7,-3.09263e-7},
+        {1.17182e-2,1.62222e-2,-5.90397e-5,-1.05585e-4,4.17873e-4,9.13843e-4},
+        {-2.65955e-1,-7.29531e-1,-4.99796e-1,2.83507e-4,-9.09042e-4,-2.20244e-3},
+        {-1.82348e-4,-8.86355e-5,-1.90554e-4,-2.49708e-4,6.35004e-4,1.73523e-3},
+        {4.70966e-5,-4.09705e-6,3.75218e-5,8.05645e-5,-1.90534e-4,-5.42847e-4}
+        }},
+        //He.......................................
+        CoeffMat{{
+        {1.e0,3.76476e-8,-3.05313e-7,-3.27422e-7,2.44235e-6,4.08754e-6},
+        {2.35767e-2,3.24642e-2,-6.37269e-4,-7.6916e-4,5.28004e-3,9.45642e-3},
+        {-2.73743e-1,-7.40767e-1,-4.98195e-1,1.74337e-3,-1.25798e-2,-2.24046e-2},
+        {-7.79128e-4,-4.14495e-4,-1.62657e-3,-1.37286e-3,1.04319e-2,1.83488e-2},
+        {2.02855e-4,1.94598e-6,4.30102e-4,4.3218e-4,-3.31526e-3,-5.81788e-3}
+        }},
+        //Li..................................................
+        CoeffMat{{
+        {1.e0,7.00357e-8,-3.15076e-7,-4.24915e-7,2.45516e-6,4.90187e-6},
+        {3.55657e-2,4.87956e-2,-1.95525e-3,-2.7866e-3,1.6549e-2,3.11496e-2},
+        {-2.81171e-1,-7.52015e-1,-4.95329e-1,5.83548e-3,-3.3983e-2,-6.55379e-2},
+        {-1.83452e-3,-8.12746e-4,-3.84675e-3,-4.44467e-3,2.55871e-2,4.99483e-2},
+        {4.79031e-4,-3.89615e-5,1.01022e-3,1.39133e-3,-7.99398e-3,-1.56366e-2}
+        }},
+        //Be..................................................
+        CoeffMat{{
+        {1.e0,7.58881e-8,4.705e-8,2.48041e-7,-2.06053e-6,-1.97319e-6},
+        {4.76788e-2,6.522e-2,-4.54331e-3,-6.50318e-3,3.76564e-2,7.17176e-2},
+        {-2.88203e-1,-7.63217e-1,-4.90337e-1,1.22839e-2,-6.86398e-2,-1.35769e-1},
+        {-3.37733e-3,-1.36514e-3,-7.51614e-3,-8.78592e-3,4.78572e-2,9.69021e-2},
+        {8.81822e-4,-1.02577e-4,1.99797e-3,2.72661e-3,-1.48296e-2,-3.0106e-2}
+        }},
+        //B....................................................
+        CoeffMat{{
+        {9.99999e-1,7.91498e-8,1.84164e-6,2.68534e-6,-1.8163e-5,-2.69021e-5},
+        {5.98818e-2,8.17654e-2,-7.70811e-3,-1.12378e-2,6.38329e-2,1.25339e-1},
+        {-2.94716e-1,-7.74405e-1,-4.8622e-1,1.77367e-2,-9.46825e-2,-2.01789e-1},
+        {-5.52375e-3,-2.05348e-3,-9.44915e-3,-1.08135e-2,5.41024e-2,1.25257e-1},
+        {1.44555e-3,-1.99404e-4,2.36742e-3,3.29655e-3,-1.64122e-2,-3.8375e-2}
+        }},
+        //C...................................................
+        CoeffMat{{
+        {9.99999e-1,7.68158e-8,5.18185e-6,7.34245e-6,-4.9478e-5,-7.71923e-5},
+        {7.21461e-2,9.84618e-2,-1.06535e-2,-1.62358e-2,8.59238e-2,1.78727e-1},
+        {-3.00622e-1,-7.85616e-1,-4.85735e-1,1.91563e-2,-8.10204e-2,-2.15386e-1},
+        {-8.34809e-3,-2.85241e-3,-7.03252e-3,-7.56786e-3,1.44975e-2,8.79093e-2},
+        {2.18964e-3,-3.42022e-4,1.3293e-3,2.20108e-3,-3.57927e-3,-2.5928e-2}
+        }},
+        //N....................................................
+        CoeffMat{{
+        {9.99999e-1,8.36312e-8,1.09116e-5,1.47812e-5,-1.02733e-4,-1.62724e-4},
+        {8.44142e-2,1.1531e-1,-1.1723e-2,-1.94732e-2,8.92604e-2,2.09303e-1},
+        {-3.05743e-1,-7.96809e-1,-4.93957e-1,1.01607e-2,1.67761e-2,-1.05909e-1},
+        {-1.2009e-2,-3.80678e-3,4.51195e-3,6.93472e-3,-1.12405e-1,-8.15484e-2},
+        {3.16048e-3,-5.22237e-4,-2.58261e-3,-2.38303e-3,3.63393e-2,2.75127e-2}
+        }},
+        //O...................................................
+        CoeffMat{{
+        {9.99998e-1,1.57323e-8,1.77595e-5,2.56082e-5,-1.67537e-4,-2.73755e-4},
+        {9.66438e-2,1.32264e-1,-9.53841e-3,-1.83707e-2,6.01664e-2,1.93357e-1},
+        {-3.09969e-1,-8.0779e-1,-5.14392e-1,-1.67153e-2,2.3387e-1,1.916e-1},
+        {-1.65906e-2,-5.11585e-3,2.80424e-2,3.94663e-2,-3.5572e-1,-4.39251e-1},
+        {4.37866e-3,-6.81795e-4,-1.0152e-2,-1.24875e-2,1.11484e-1,1.38105e-1}
+        }},
+        //F.................................................
+        CoeffMat{{
+        {9.99997e-1,-8.06132e-8,2.49797e-5,3.8512e-5,-2.37451e-4,-3.99607e-4},
+        {1.08782e-1,1.49306e-1,-2.50975e-3,-1.05471e-2,-1.64831e-2,1.05733e-1},
+        {-3.13165e-1,-8.18489e-1,-5.50832e-1,-6.74447e-2,6.06357e-1,7.39717e-1},
+        {-2.21976e-2,-6.84023e-3,6.65411e-2,9.48702e-2,-7.43989e-1,-1.03582e0},
+        {5.87088e-3,-8.1048e-4,-2.21731e-2,-2.94422e-2,2.2954e-1,3.19669e-1}
+        }},
+        //Ne.................................................
+        CoeffMat{{
+        {9.99997e-1,-1.87404e-7,3.10276e-5,5.20007e-5,-2.98132e-4,-5.19259e-4},
+        {1.20783e-1,1.66407e-1,1.06608e-2,6.48772e-3,-1.53031e-1,-7.59354e-2},
+        {-3.15222e-1,-8.28793e-1,-6.0574e-1,-1.47812e-1,1.1576e0,1.58565e0},
+        {-2.89055e-2,-9.08096e-3,1.21467e-1,1.77575e-1,-1.2911e0,-1.90333e0},
+        {7.65342e-3,-8.85417e-4,-3.89092e-2,-5.4404e-2,3.93087e-1,5.79439e-1}
+        }},
+        //Na.................................................
+        CoeffMat{{
+        {9.99996e-1,-2.44548e-7,3.31019e-5,6.29483e-5,-3.24667e-4,-5.95527e-4},
+        {1.32615e-1,1.83566e-1,3.04158e-2,3.40925e-2,-3.54681e-1,-3.63044e-1},
+        {-3.16092e-1,-8.38704e-1,-6.78558e-1,-2.59346e-1,1.88547e0,2.73632e0},
+        {-3.67233e-2,-1.18139e-2,1.91089e-1,2.87408e-1,-1.98397e0,-3.03075e0},
+        {9.72033e-3,-9.2638e-4,-5.95654e-2,-8.69829e-2,5.95744e-1,9.10242e-1}
+        }},
+        //Mg.................................................
+        CoeffMat{{
+        {9.99995e-1,-2.12227e-7,2.95645e-5,6.92848e-5,-3.02153e-4,-6.05145e-4},
+        {1.44258e-1,2.00775e-1,5.67845e-2,7.35166e-2,-6.22861e-1,-7.62213e-1},
+        {-3.15754e-1,-8.48196e-1,-7.67318e-1,-4.02984e-1,2.77477e0,4.18114e0},
+        {-4.56307e-2,-1.50425e-2,2.72232e-1,4.23528e-1,-2.79606e0,-4.38863e0},
+        {1.2056e-2,-9.44637e-4,-8.28738e-2,-1.26564e-1,8.26726e-1,1.29882e0}
+        }},
+        //Al.....................................................
+        CoeffMat{{
+        {9.99995e-1,-4.03407e-8,1.86047e-5,6.85201e-5,-2.14503e-4,-5.22528e-4},
+        {1.55704e-1,2.18048e-1,8.88994e-2,1.24878e-1,-9.51331e-1,-1.26824e0},
+        {-3.14244e-1,-8.57322e-1,-8.6719e-1,-5.75787e-1,3.78571e0,5.87052e0},
+        {-5.55526e-2,-1.86861e-2,3.5886e-1,5.81094e-1,-3.67623e0,-5.908e0},
+        {1.46269e-2,-9.79742e-4,-1.06652e-1,-1.71226e-1,1.06737e0,1.71918e0}
+        }},
+        //Si.....................................................
+        CoeffMat{{
+        {9.99994e-1,3.00267e-7,-1.1184e-6,5.88256e-5,-4.78456e-5,-3.25731e-4},
+        {1.6696e-1,2.35405e-1,1.25215e-1,1.87646e-1,-1.32685e0,-1.86549e0},
+        {-3.1163e-1,-8.66152e-1,-9.71254e-1,-7.72715e-1,4.85654e0,7.7215e0},
+        {-6.63778e-2,-2.26481e-2,4.42898e-1,7.53182e-1,-4.55172e0,-7.4867e0},
+        {1.73883e-2,-1.07669e-3,-1.28075e-1,-2.18389e-1,1.29217e0,2.13475e0}
+        }},
+        //P.....................................................
+        CoeffMat{{
+        {9.99994e-1,8.94829e-7,-2.98434e-5,3.82193e-5,2.00584e-4,-6.40482e-6},
+        {1.78039e-1,2.52912e-1,1.63761e-1,2.60132e-1,-1.73287e0,-2.53185e0},
+        {-3.08007e-1,-8.74905e-1,-1.07146e0,-9.85062e-1,5.91697e0,9.63265e0},
+        {-7.79747e-2,-2.66797e-2,5.15288e-1,9.29261e-1,-5.34252e0,-9.00574e0},
+        {2.02892e-2,-1.33011e-3,-1.44039e-1,-2.6433e-1,1.4736e0,2.50398e0}
+        }},
+        //S....................................................
+        CoeffMat{{
+        {9.99994e-1,1.75397e-6,-6.7331e-5,6.29524e-6,5.29623e-4,4.35288e-4},
+        {1.88968e-1,2.70612e-1,2.01975e-1,3.40574e-1,-2.14737e0,-3.23836e0},
+        {-3.03499e-1,-8.83717e-1,-1.15816e0,-1.20414e0,6.88176e0,1.14841e1},
+        {-9.01806e-2,-3.06202e-2,5.65581e-1,1.09902e0,-5.95552e0,-1.03302e1},
+        {2.32694e-2,-1.80614e-3,-1.51041e-1,-3.05449e-1,1.58037e0,2.78083e0}
+        }},
+        //Cl....................................................
+        CoeffMat{{
+        {9.99994e-1,3.07931e-6,-1.11876e-4,-4.10164e-5,9.17095e-4,9.80145e-4},
+        {1.99765e-1,2.88611e-1,2.37501e-1,4.25803e-1,-2.55105e0,-3.95585e0},
+        {-2.98206e-1,-8.9293e-1,-1.22279e0,-1.4169e0,7.67836e0,1.31601e1},
+        {-1.02865e-1,-3.40967e-2,5.84677e-1,1.24786e0,-6.31301e0,-1.13328e1},
+        {2.628e-2,-2.63995e-3,-1.46076e-1,-3.36795e-1,1.58677e0,2.92251e0}
+        }},
+        //Ar....................................................
+        CoeffMat{{
+        {9.99993e-1,4.49776e-6,-1.65136e-4,-9.76754e-5,1.39664e-3,1.66293e-3},
+        {2.10469e-1,3.06924e-1,2.66793e-1,5.13797e-1,-2.90958e0,-4.63816e0},
+        {-2.92294e-1,-9.0256e-1,-1.25307e0,-1.6147e0,8.18574e0,1.44912e1},
+        {-1.15831e-1,-3.70891e-2,5.59807e-1,1.36619e0,-6.28824e0,-1.18327e1},
+        {2.92513e-2,-3.84903e-3,-1.24976e-1,-3.55149e-1,1.45127e0,2.86925e0}
+        }},
+        //K.................................................
+        CoeffMat{{
+        {9.99993e-1,6.01488e-6,-2.22125e-4,-1.61774e-4,1.92058e-3,2.41975e-3},
+        {2.21091e-1,3.25648e-1,2.87732e-1,6.01632e-1,-3.20436e0,-5.25724e0},
+        {-2.85814e-1,-9.12889e-1,-1.24213e0,-1.78635e0,8.34196e0,1.53776e1},
+        {-1.29005e-1,-3.92986e-2,4.84255e-1,1.44206e0,-5.81999e0,-1.17275e1},
+        {3.21555e-2,-5.54272e-3,-8.56572e-2,-3.56589e-1,1.1548e0,2.58829e0}
+        }},
+        //Ca.................................................
+        CoeffMat{{
+        {9.99993e-1,8.01467e-6,-2.79242e-4,-2.3682e-4,2.45459e-3,3.21683e-3},
+        {2.31651e-1,3.44948e-1,2.9782e-1,6.85187e-1,-3.41294e0,-5.77715e0},
+        {-2.78858e-1,-9.24428e-1,-1.18215e0,-1.91691e0,8.07489e0,1.56969e1},
+        {-1.42276e-1,-4.01888e-2,3.50466e-1,1.45983e0,-4.83806e0,-1.08936e1},
+        {3.49529e-2,-7.90933e-3,-2.58002e-2,-3.36028e-1,6.7574e-1,2.04052e0}
+        }},
+        //Sc.................................................
+        CoeffMat{{
+        {9.99992e-1,1.04277e-5,-3.35126e-4,-3.21042e-4,2.98507e-3,4.03325e-3},
+        {2.42172e-1,3.64954e-1,2.94606e-1,7.60693e-1,-3.51409e0,-6.1646e0},
+        {-2.71512e-1,-9.37543e-1,-1.0657e0,-1.99328e0,7.31863e0,1.53396e1},
+        {-1.5554e-1,-3.93862e-2,1.51477e-1,1.40614e0,-3.28024e0,-9.22338e0},
+        {3.76066e-2,-1.10812e-2,5.66831e-2,-2.8921e-1,-4.60274e-3,1.19273e0}
+        }},
+        //Ti.................................................
+        CoeffMat{{
+        {9.99992e-1,1.30838e-5,-3.8407e-4,-4.09294e-4,3.47025e-3,4.81071e-3},
+        {2.52646e-1,3.85718e-1,2.7633e-1,8.25665e-1,-3.48888e0,-6.39017e0},
+        {-2.63758e-1,-9.52326e-1,-8.88179e-1,-2.0072e0,6.0196e0,1.42157e1},
+        {-1.68806e-1,-3.68095e-2,-1.16429e-1,1.27309e0,-1.09991e0,-6.63409e0},
+        {4.01184e-2,-1.50938e-2,1.6274e-1,-2.1376e-1,-8.99142e-1,2.08129e-2}
+        }},
+        //V.................................................
+        CoeffMat{{
+        {9.99991e-1,1.59363e-5,-4.27315e-4,-5.01461e-4,3.91365e-3,5.55096e-3},
+        {2.63096e-1,4.07357e-1,2.40649e-1,8.7645e-1,-3.31641e0,-6.42069e0},
+        {-2.55682e-1,-9.6909e-1,-6.43149e-1,-1.94678e0,4.11915e0,1.22255e1},
+        {-1.81974e-1,-3.21462e-2,-4.58817e-1,1.04913e0,1.75388e0,-3.03434e0},
+        {4.24541e-2,-2.00595e-2,2.93916e-1,-1.06138e-1,-2.02203e0,-1.50181e0}
+        }},
+        //Cr.................................................
+        CoeffMat{{
+        {9.9999e-1,1.8895e-5,-4.59994e-4,-5.93663e-4,4.27684e-3,6.2009e-3},
+        {2.73504e-1,4.2999e-1,1.86473e-1,9.09921e-1,-2.98441e0,-6.23344e0},
+        {-2.47225e-1,-9.88118e-1,-3.28759e-1,-1.80252e0,1.5909e0,9.30968e0},
+        {-1.951e-1,-2.51242e-2,-8.76244e-1,7.25592e-1,5.2965e0,1.62177e0},
+        {4.46307e-2,-2.60754e-2,4.50054e-1,3.61638e-2,-3.37524e0,-3.38601e0}
+        }},
+        //Mn.................................................
+        CoeffMat{{
+        {9.99989e-1,2.18906e-5,-4.79199e-4,-6.83164e-4,4.53467e-3,6.72485e-3},
+        {2.83854e-1,4.53722e-1,1.12877e-1,9.23203e-1,-2.48247e0,-5.80855e0},
+        {-2.38338e-1,-1.00965e0,5.61453e-2,-1.56605e0,-1.58353e0,5.42138e0},
+        {-2.08226e-1,-1.55138e-2,-1.36846e0,2.95168e-1,9.53392e0,7.3655e0},
+        {4.66619e-2,-3.32255e-2,6.30711e-1,2.15167e-1,-4.95748e0,-5.63744e0}
+        }},
+        //Fe...................................................
+        CoeffMat{{
+        {9.99987e-1,2.482e-5,-4.82895e-4,-7.67488e-4,4.669e-3,7.09581e-3},
+        {2.94123e-1,4.78653e-1,1.93256e-2,9.13569e-1,-1.80354e0,-5.1304e0},
+        {-2.28945e-1,-1.0339e0,5.1121e-1,-1.22996e0,-5.40939e0,5.3149e-1},
+        {-2.21426e-1,-3.12679e-3,-1.93353e0,-2.48179e-1,1.44576e1,1.42077e1},
+        {4.85718e-2,-4.15796e-2,8.34879e-1,4.32418e-1,-6.76244e0,-8.25464e0}
+        }},
+        //Co..................................................
+        CoeffMat{{
+        {9.99985e-1,2.76168e-5,-4.67522e-4,-8.43978e-4,4.65008e-3,7.27476e-3},
+        {3.0428e-1,5.04856e-1,-9.42256e-2,8.78905e-1,-9.44256e-1,-4.18847e0},
+        {-2.18939e-1,-1.06098e0,1.03431e0,-7.89123e-1,-9.87815e0,-5.36997e0},
+        {-2.34805e-1,1.21281e-2,-2.56765e0,-9.08015e-1,2.00436e1,2.21379e1},
+        {5.03942e-2,-5.11768e-2,1.0609e0,6.88631e-1,-8.77867e0,-1.12289e1}
+        }},
+        //Ni..................................................
+        CoeffMat{{
+        {9.99982e-1,3.00792e-5,-4.33447e-4,-9.09366e-4,4.47786e-3,7.25072e-3},
+        {3.14283e-1,5.32444e-1,-2.27528e-1,8.16951e-1,9.53704e-2,-2.9773e0},
+        {-2.08186e-1,-1.09112e0,1.62237e0,-2.38394e-1,-1.49699e1,-1.22743e1},
+        {-2.48493e-1,3.04543e-2,-3.26601e0,-1.6876e0,2.62568e1,3.11259e1},
+        {5.21709e-2,-6.20908e-2,1.30683e0,9.84353e-1,-1.09909e1,-1.45449e1}
+        }},
+        //Cu...................................................
+        CoeffMat{{
+        {9.99979e-1,3.24569e-5,-3.76717e-4,-9.64e-4,4.11997e-3,6.98761e-3},
+        {3.24082e-1,5.61534e-1,-3.79993e-1,7.25313e-1,1.31342e0,-1.49163e0},
+        {-1.96521e-1,-1.12457e0,2.27091e0,4.27678e-1,-2.06558e1,-2.0169e1},
+        {-2.62655e-1,5.20812e-2,-4.02224e0,-2.59048e0,3.30508e1,4.1135e1},
+        {5.39556e-2,-7.44077e-2,1.57015e0,1.32022e0,-1.33802e1,-1.81848e1}
+        }},
+        //Zn...................................................
+        CoeffMat{{
+        {9.99976e-1,3.39628e-5,-2.98845e-4,-9.99651e-4,3.58523e-3,6.47782e-3},
+        {3.33647e-1,5.92068e-1,-5.5102e-1,6.0363e-1,2.70712e0,2.67853e-1},
+        {-1.83849e-1,-1.16095e0,2.97559e0,1.20728e0,-2.69053e1,-2.90214e1},
+        {-2.77375e-1,7.65838e-2,-4.83023e0,-3.61238e0,4.03786e1,5.21084e1},
+        {5.57745e-2,-8.79952e-2,1.84848e0,1.69425e0,-1.59274e1,-2.21242e1}
+        }},
+        //Ga...................................................
+        CoeffMat{{
+        {9.99972e-1,3.48473e-5,-1.97828e-4,-1.01659e-3,2.85509e-3,5.69725e-3},
+        {3.42904e-1,6.24164e-1,-7.39023e-1,4.50306e-1,4.26553e0,2.29299e0},
+        {-1.6994e-1,-1.20051e0,3.72868e0,2.10276e0,-3.36594e1,-3.87726e1},
+        {-2.92882e-1,1.04194e-1,-5.68033e0,-4.75339e0,4.81633e1,6.39614e1},
+        {5.77008e-2,-1.02941e-1,2.13826e0,2.10589e0,-1.86034e1,-2.63294e1}
+        }},
+        //Ge.................................................
+        CoeffMat{{
+        {9.99968e-1,3.47804e-5,-7.11898e-5,-1.01028e-3,1.90919e-3,4.61426e-3},
+        {3.51793e-1,6.57815e-1,-9.42175e-1,2.6494e-1,5.97606e0,4.57222e0},
+        {-1.54595e-1,-1.24305e0,4.5217e0,3.11205e0,-4.08539e1,-4.93518e1},
+        {-3.09367e-1,1.34661e-1,-6.56214e0,-6.00882e0,5.63228e1,7.65965e1},
+        {5.97948e-2,-1.19174e-1,2.4357e0,2.553e0,-2.13779e1,-3.07628e1}
+        }},
+        //As.................................................
+        CoeffMat{{
+        {9.99963e-1,3.37519e-5,8.13037e-5,-9.78638e-4,7.41412e-4,3.21498e-3},
+        {3.60246e-1,6.93065e-1,-1.1588e0,4.68519e-2,7.82662e0,7.09456e0},
+        {-1.37615e-1,-1.28855e0,5.34673e0,4.23402e0,-4.84263e1,-6.06897e1},
+        {-3.27017e-1,1.67941e-1,-7.46593e0,-7.37492e0,6.47773e1,8.99181e1},
+        {6.21158e-2,-1.36692e-1,2.73726e0,3.03374e0,-2.42209e1,-3.53873e1}
+        }},
+        //Se.................................................
+        CoeffMat{{
+        {9.99958e-1,3.14983e-5,2.59741e-4,-9.19008e-4,-6.49202e-4,1.49008e-3},
+        {3.68196e-1,7.29888e-1,-1.38664e0,-2.03807e-1,9.80062e0,9.84155e0},
+        {-1.18788e-1,-1.33674e0,6.194e0,5.46446e0,-5.62994e1,-7.26915e1},
+        {-3.46033e-1,2.03729e-1,-8.38012e0,-8.84466e0,7.34319e1,1.03805e2},
+        {6.47255e-2,-1.55409e-1,3.03879e0,3.54516e0,-2.7098e1,-4.01572e1}
+        }},
+        //Br.................................................
+        CoeffMat{{
+        {9.99952e-1,2.79961e-5,4.60479e-4,-8.33486e-4,-2.23214e-3,-5.16285e-4},
+        {3.7558e-1,7.68292e-1,-1.62392e0,-4.87674e-1,1.18854e1,1.28025e1},
+        {-9.79355e-2,-1.38749e0,7.0556e0,6.80198e0,-6.44113e1,-8.52915e1},
+        {-3.66572e-1,2.41863e-1,-9.29524e0,-1.04141e1,8.22093e1,1.18166e2},
+        {6.76714e-2,-1.75289e-1,3.33691e0,4.08538e0,-2.99809e1,-4.50381e1}
+        }},
+        //Kr.................................................
+        CoeffMat{{
+        {9.99947e-1,2.2952e-5,6.82639e-4,-7.17139e-4,-4.00522e-3,-2.81601e-3},
+        {3.82332e-1,8.08194e-1,-1.86848e0,-8.03415e-1,1.4064e1,1.5956e1},
+        {-7.48735e-2,-1.44031e0,7.92233e0,8.2382e0,-7.26858e1,-9.83865e1},
+        {-3.88797e-1,2.81837e-1,-1.02005e1,-1.20716e1,9.10175e1,1.32873e2},
+        {7.10017e-2,-1.96182e-1,3.6278e0,4.65e0,-3.28365e1,-4.99823e1}
+        }},
+        //Rb.................................................
+        CoeffMat{{
+        {9.99941e-1,1.63607e-5,9.28242e-4,-5.69257e-4,-5.98245e-3,-5.42476e-3},
+        {3.88363e-1,8.49548e-1,-2.11701e0,-1.15003e0,1.63119e1,1.92737e1},
+        {-4.93364e-2,-1.49491e0,8.78129e0,9.76596e0,-8.10212e1,-1.11851e2},
+        {-4.12953e-1,3.23322e-1,-1.10814e1,-1.38073e1,9.97386e1,1.47775e2},
+        {7.47911e-2,-2.18001e-1,3.90642e0,5.23509e0,-3.56231e1,-5.49352e1}
+        }},
+        //Sr.................................................
+        CoeffMat{{
+        {9.99935e-1,8.3152e-6,1.19244e-3,-3.95258e-4,-8.12525e-3,-8.28836e-3},
+        {3.93603e-1,8.92358e-1,-2.36688e0,-1.52786e0,1.86071e1,2.27314e1},
+        {-2.11345e-2,-1.55112e0,9.62205e0,1.13827e1,-8.93266e1,-1.25574e2},
+        {-4.39199e-1,3.66161e-1,-1.19262e1,-1.56159e1,1.08267e2,1.62736e2},
+        {7.90853e-2,-2.40719e-1,4.16872e0,5.83839e0,-3.8304e1,-5.98484e1}
+        }},
+        //Y.................................................
+        CoeffMat{{
+        {9.99929e-1,-1.65608e-6,1.47476e-3,-1.8625e-4,-1.04363e-2,-1.14318e-2},
+        {3.97989e-1,9.3643e-1,-2.61569e0,-1.93387e0,2.09305e1,2.63021e1},
+        {9.88744e-3,-1.60813e0,1.04351e1,1.3074e1,-9.75207e1,-1.39437e2},
+        {-4.67657e-1,4.09494e-1,-1.2724e1,-1.74797e1,1.16508e2,1.77614e2},
+        {8.39169e-2,-2.64077e-1,4.41092e0,6.45344e0,-4.08455e1,-6.46698e1}
+        }},
+        //Zr.................................................
+        CoeffMat{{
+        {9.99922e-1,-1.37624e-5,1.77335e-3,5.994e-5,-1.29013e-2,-1.48443e-2},
+        {4.0145e-1,9.816e-1,-2.86053e0,-2.36562e0,2.32591e1,2.99551e1},
+        {4.3915e-2,-1.66522e0,1.12093e1,1.48279e1,-1.05512e2,-1.5331e2},
+        {-4.98475e-1,4.52607e-1,-1.34628e1,-1.93836e1,1.24357e2,1.92256e2},
+        {8.93258e-2,-2.87869e-1,4.62891e0,7.07473e0,-4.32116e1,-6.93461e1}
+        }},
+        //Nb.................................................
+        CoeffMat{{
+        {9.99916e-1,-2.78975e-5,2.08753e-3,3.41607e-4,-1.55141e-2,-1.85156e-2},
+        {4.03889e-1,1.02783e0,-3.09762e0,-2.82183e0,2.55629e1,3.36548e1},
+        {8.12122e-2,-1.7221e0,1.1931e1,1.66362e1,-1.13185e2,-1.67048e2},
+        {-5.3188e-1,4.95236e-1,-1.41275e1,-2.13166e1,1.31687e2,2.06495e2},
+        {9.53781e-2,-3.12041e-1,4.81765e0,7.69813e0,-4.53587e1,-7.38182e1}
+        }},
+        //Mo.................................................
+        CoeffMat{{
+        {9.9991e-1,-4.3988e-5,2.41439e-3,6.54636e-4,-1.82489e-2,-2.24062e-2},
+        {4.05249e-1,1.07495e0,-3.32445e0,-3.30073e0,2.78221e1,3.7376e1},
+        {1.21896e-1,-1.77813e0,1.25907e1,1.84893e1,-1.20461e2,-1.80542e2},
+        {-5.67942e-1,5.36741e-1,-1.4708e1,-2.32664e1,1.38408e2,2.20203e2},
+        {1.02086e-1,-3.36419e-1,4.97373e0,8.31905e0,-4.72562e1,-7.80409e1}
+        }},
+        //Tc.................................................
+        CoeffMat{{
+        {9.99904e-1,-6.22073e-5,2.75066e-3,1.00063e-3,-2.10817e-2,-2.64933e-2},
+        {4.05472e-1,1.12279e0,-3.53825e0,-3.79964e0,3.00132e1,4.10871e1},
+        {1.66089e-1,-1.83255e0,1.31785e1,2.03743e1,-1.27252e2,-1.93664e2},
+        {-6.06729e-1,5.76392e-1,-1.51936e1,-2.52171e1,1.44423e2,2.33232e2},
+        {1.09463e-1,-3.60803e-1,5.09358e0,8.93178e0,-4.88706e1,-8.19632e1}
+        }},
+        //Ru.................................................
+        CoeffMat{{
+        {9.99898e-1,-8.26232e-5,3.09353e-3,1.37924e-3,-2.39902e-2,-3.07502e-2},
+        {4.04498e-1,1.17115e0,-3.73628e0,-4.3158e0,3.21142e1,4.4758e1},
+        {2.13904e-1,-1.88458e0,1.36848e1,2.22782e1,-1.33471e2,-2.06293e2},
+        {-6.48301e-1,6.13439e-1,-1.55742e1,-2.71531e1,1.49638e2,2.45444e2},
+        {1.17516e-1,-3.8499e-1,5.17383e0,9.5307e0,-5.01709e1,-8.55368e1}
+        }},
+        //Rh.................................................
+        CoeffMat{{
+        {9.99893e-1,-1.05293e-4,3.43932e-3,1.78981e-3,-2.69462e-2,-3.51436e-2},
+        {4.02272e-1,1.21982e0,-3.91575e0,-4.84621e0,3.41013e1,4.83563e1},
+        {2.65438e-1,-1.93341e0,1.40998e1,2.41876e1,-1.39034e2,-2.183e2},
+        {-6.92691e-1,6.47135e-1,-1.58398e1,-2.90581e1,1.5396e2,2.56697e2},
+        {1.26243e-1,-4.08783e-1,5.21125e0,1.011e1,-5.11262e1,-8.8713e1}
+        }},
+        //Pd.................................................
+        CoeffMat{{
+        {9.99888e-1,-1.30188e-4,3.78436e-3,2.22971e-3,-2.99185e-2,-3.96315e-2},
+        {3.9875e-1,1.26855e0,-4.07427e0,-5.38796e0,3.5955e1,5.18546e1},
+        {3.20748e-1,-1.97817e0,1.44153e1,2.6089e1,-1.43866e2,-2.29578e2},
+        {-7.39891e-1,6.76679e-1,-1.59819e1,-3.09162e1,1.57312e2,2.66867e2},
+        {1.3563e-1,-4.31972e-1,5.2031e0,1.06642e1,-5.17104e1,-9.14493e1}
+        }},
+        //Ag.................................................
+        CoeffMat{{
+        {9.99884e-1,-1.57459e-4,4.12052e-3,2.69984e-3,-3.28468e-2,-4.41512e-2},
+        {3.93891e-1,1.3171e0,-4.20944e0,-5.93691e0,3.76535e1,5.52188e1},
+        {3.79857e-1,-2.01791e0,1.46238e1,2.79653e1,-1.47893e2,-2.39997e2},
+        {-7.89852e-1,7.01196e-1,-1.59929e1,-3.27077e1,1.59613e2,2.75813e2},
+        {1.45646e-1,-4.54328e-1,5.14701e0,1.11863e1,-5.18977e1,-9.36987e1}
+        }},
+        //Cd.................................................
+        CoeffMat{{
+        {9.99881e-1,-1.86816e-4,4.45491e-3,3.19574e-3,-3.57812e-2,-4.87457e-2},
+        {3.87612e-1,1.36524e0,-4.31686e0,-6.49005e0,3.91612e1,5.84032e1},
+        {4.4294e-1,-2.05189e0,1.47103e1,2.98033e1,-1.50988e2,-2.49389e2},
+        {-8.42683e-1,7.20045e-1,-1.58579e1,-3.44169e1,1.60733e2,2.83354e2},
+        {1.56312e-1,-4.75705e-1,5.0381e0,1.16708e1,-5.16454e1,-9.53999e1}
+        }},
+        //In.................................................
+        CoeffMat{{
+        {9.99878e-1,-2.1848e-4,4.78376e-3,3.71897e-3,-3.86926e-2,-5.33866e-2},
+        {3.79879e-1,1.41265e0,-4.39391e0,-7.04282e0,4.0455e1,6.13721e1},
+        {5.10001e-1,-2.07896e0,1.46666e1,3.15843e1,-1.53072e2,-2.57624e2},
+        {-8.98306e-1,7.32213e-1,-1.55686e1,-3.60229e1,1.60591e2,2.89348e2},
+        {1.67591e-1,-4.95837e-1,4.87389e0,1.21106e1,-5.09274e1,-9.6506e1}
+        }},
+        //Sn................................................
+        CoeffMat{{
+        {9.99876e-1,-2.52173e-4,5.09558e-3,4.26518e-3,-4.14939e-2,-5.79712e-2},
+        {3.70689e-1,1.45908e0,-4.43983e0,-7.59192e0,4.15261e1,6.4108e1},
+        {5.80924e-1,-2.09831e0,1.44911e1,3.32945e1,-1.54118e2,-2.64634e2},
+        {-9.56518e-1,7.37009e-1,-1.51243e1,-3.75098e1,1.59159e2,2.93723e2},
+        {1.79397e-1,-5.14573e-1,4.65426e0,1.25003e1,-4.97362e1,-9.69932e1}
+        }},
+        //Sb.................................................
+        CoeffMat{{
+        {9.99874e-1,-2.87917e-4,5.39066e-3,4.83313e-3,-4.41852e-2,-6.24953e-2},
+        {3.60001e-1,1.5042e0,-4.4515e0,-8.1327e0,4.23476e1,6.65709e1},
+        {6.55716e-1,-2.10882e0,1.41741e1,3.49157e1,-1.54034e2,-2.70276e2},
+        {-1.01724e0,7.33501e-1,-1.45156e1,-3.88572e1,1.56348e2,2.96328e2},
+        {1.9169e-1,-5.3169e-1,4.37639e0,1.28328e1,-4.80435e1,-9.68125e1}
+        }},
+        //Te.................................................
+        CoeffMat{{
+        {9.99874e-1,-3.25847e-4,5.67083e-3,5.42945e-3,-4.67877e-2,-6.701e-2},
+        {3.47783e-1,1.54767e0,-4.42606e0,-8.65995e0,4.28934e1,6.87183e1},
+        {7.34353e-1,-2.10941e0,1.3707e1,3.64273e1,-1.52736e2,-2.74401e2},
+        {-1.08036e0,7.20771e-1,-1.37344e1,-4.00427e1,1.52073e2,2.97007e2},
+        {2.04417e-1,-5.46972e-1,4.03784e0,1.31007e1,-4.58224e1,-9.59123e1}
+        }},
+        //I.................................................
+        CoeffMat{{
+        {9.99875e-1,-3.65406e-4,5.92115e-3,6.03332e-3,-4.91674e-2,-7.12951e-2},
+        {3.34062e-1,1.58924e0,-4.36361e0,-9.17241e0,4.31638e1,7.05522e1},
+        {8.16604e-1,-2.09933e0,1.30915e1,3.78235e1,-1.50231e2,-2.77015e2},
+        {-1.14554e0,6.98298e-1,-1.2784e1,-4.10591e1,1.46348e2,2.9577e2},
+        {2.17451e-1,-5.60344e-1,3.63994e0,1.33015e1,-4.30802e1,-9.42981e1}
+        }},
+        //Xe..................................................
+        CoeffMat{{
+        {9.99877e-1,-4.06637e-4,6.14278e-3,6.64937e-3,-5.13391e-2,-7.53881e-2},
+        {3.18822e-1,1.62854e0,-4.26169e0,-9.66459e0,4.31357e1,7.20331e1},
+        {9.02373e-1,-2.07743e0,1.23209e1,3.90832e1,-1.46445e2,-2.77984e2},
+        {-1.21259e0,6.65174e-1,-1.16583e1,-4.18839e1,1.39104e2,2.92479e2},
+        {2.3071e-1,-5.71608e-1,3.18104e0,1.34277e1,-3.97963e1,-9.19255e1}
+        }},
+        //Cs..................................................
+        CoeffMat{{
+        {9.99881e-1,-4.49175e-4,6.32631e-3,7.27184e-3,-5.32297e-2,-7.91972e-2},
+        {3.02086e-1,1.66529e0,-4.11996e0,-1.0133e1,4.28031e1,7.31464e1},
+        {9.91429e-1,-2.04292e0,1.13959e1,4.01931e1,-1.41369e2,-2.7726e2},
+        {-1.28117e0,6.2087e-1,-1.03596e1,-4.25024e1,1.30338e2,2.8709e2},
+        {2.44065e-1,-5.80702e-1,2.66217e0,1.34745e1,-3.59721e1,-8.87824e1}
+        }},
+        //Ba..................................................
+        CoeffMat{{
+        {9.99886e-1,-4.93136e-4,6.47914e-3,7.90029e-3,-5.48909e-2,-8.27736e-2},
+        {2.8384e-1,1.69905e0,-3.93534e0,-1.05724e1,4.21392e1,7.38521e1},
+        {1.08366e0,-1.99451e0,1.03076e1,4.1134e1,-1.34917e2,-2.7471e2},
+        {-1.35107e0,5.64397e-1,-8.87993e0,-4.28945e1,1.19972e2,2.7947e2},
+        {2.57431e-1,-5.87417e-1,2.08111e0,1.34353e1,-3.15844e1,-8.48268e1}
+        }},
+        //La..................................................
+        CoeffMat{{
+        {9.99892e-1,-5.38122e-4,6.59206e-3,8.52848e-3,-5.62508e-2,-8.60251e-2},
+        {2.6413e-1,1.72952e0,-3.70812e0,-1.09794e1,4.1143e1,7.4141e1},
+        {1.17876e0,-1.93141e0,9.05928e0,4.18928e1,-1.27099e2,-2.70309e2},
+        {-1.42185e0,4.95292e-1,-7.22427e0,-4.30461e1,1.08024e2,2.69601e2},
+        {2.70647e-1,-5.91727e-1,1.43982e0,1.33056e1,-2.66423e1,-8.00557e1}
+        }},
+        //Ce..................................................
+        CoeffMat{{
+        {9.999e-1,-5.83896e-4,6.66373e-3,9.15389e-3,-5.72988e-2,-8.89368e-2},
+        {2.42961e-1,1.75635e0,-3.43619e0,-1.13482e1,3.97928e1,7.39747e1},
+        {1.27654e0,-1.85263e0,7.64589e0,4.24492e1,-1.17853e2,-2.63933e2},
+        {-1.49322e0,4.12926e-1,-5.38909e0,-4.29363e1,9.44419e1,2.57363e2},
+        {2.83597e-1,-5.93557e-1,7.3754e-1,1.30785e1,-2.11316e1,-7.44329e1}
+        }},
+        //Pr..................................................
+        CoeffMat{{
+        {9.99909e-1,-6.301e-4,6.69295e-3,9.77077e-3,-5.8022e-2,-9.14799e-2},
+        {2.20373e-1,1.77919e0,-3.11907e0,-1.16754e1,3.80818e1,7.33384e1},
+        {1.37668e0,-1.75734e0,6.06819e0,4.27905e1,-1.07169e2,-2.55543e2},
+        {-1.56475e0,3.16885e-1,-3.37687e0,-4.25518e1,7.92269e1,2.42726e2},
+        {2.96124e-1,-5.92915e-1,-2.45072e-2,1.27501e1,-1.50562e1,-6.79514e1}
+        }},
+        //Nd..................................................
+        CoeffMat{{
+        {9.9992e-1,-6.77211e-4,6.67917e-3,1.03866e-2,-5.84208e-2,-9.36775e-2},
+        {1.96406e-1,1.79753e0,-2.7553e0,-1.19533e1,3.59929e1,7.21943e1},
+        {1.47887e0,-1.64416e0,4.32381e0,4.28906e1,-9.50034e1,-2.45021e2},
+        {-1.636e0,2.06169e-1,-1.18704e0,-4.18661e1,6.23476e1,2.25579e2},
+        {3.08067e-1,-5.89622e-1,-8.4608e-1,1.23116e1,-8.40939e0,-6.05788e1}
+        }},
+        //Pm..................................................
+        CoeffMat{{
+        {9.99932e-1,-7.23543e-4,6.61026e-3,1.09756e-2,-5.83892e-2,-9.53462e-2},
+        {1.71135e-1,1.81127e0,-2.34665e0,-1.21818e1,3.35368e1,7.05547e1},
+        {1.58269e0,-1.5131e0,2.4214e0,4.27487e1,-8.14091e1,-2.32422e2},
+        {-1.70642e0,8.13302e-2,1.16945e0,-4.0879e1,4.38728e1,2.05995e2},
+        {3.19226e-1,-5.84023e-1,-1.72313e0,1.17636e1,-1.21776e0,-5.23434e1}
+        }},
+        //Sm..................................................
+        CoeffMat{{
+        {9.99946e-1,-7.69368e-4,6.48923e-3,1.15425e-2,-5.79523e-2,-9.65284e-2},
+        {1.44613e-1,1.81998e0,-1.89196e0,-1.23548e1,3.06998e1,6.83894e1},
+        {1.68776e0,-1.36316e0,3.59801e-1,4.23443e1,-6.63559e1,-2.17658e2},
+        {-1.77552e0,-5.81387e-2,3.69184e0,-3.95701e1,2.37855e1,1.83897e2},
+        {3.2942e-1,-5.76119e-1,-2.65496e0,1.10995e1,6.52027e0,-4.32241e1}
+        }},
+        //Eu..................................................
+        CoeffMat{{
+        {9.99962e-1,-8.13654e-4,6.31217e-3,1.20763e-2,-5.70782e-2,-9.71683e-2},
+        {1.16915e-1,1.82343e0,-1.39208e0,-1.24689e1,2.74833e1,6.56914e1},
+        {1.79364e0,-1.194e0,-1.85513e0,4.16665e1,-4.9868e1,-2.00724e2},
+        {-1.84272e0,-2.12012e-1,6.37189e0,-3.79293e1,2.12742e0,1.59295e2},
+        {3.38449e-1,-5.66167e-1,-3.63838e0,1.03165e1,1.47866e1,-3.32296e1}
+        }},
+        //Gd..................................................
+        CoeffMat{{
+        {9.9998e-1,-8.57331e-4,6.07736e-3,1.25819e-2,-5.57471e-2,-9.72475e-2},
+        {8.8154e-2,1.82105e0,-8.47718e-1,-1.25186e1,2.38891e1,6.2451e1},
+        {1.89976e0,-1.00428e0,-4.21799e0,4.0697e1,-3.19703e1,-1.81604e2},
+        {-1.90731e0,-3.8108e-1,9.2019e0,-3.59386e1,-2.10594e1,1.32192e2},
+        {3.46059e-1,-5.54093e-1,-4.67039e0,9.40924e0,2.35631e1,-2.23653e1}
+        }},
+        //Tb..................................................
+        CoeffMat{{
+        {9.99999e-1,-8.99343e-4,5.7868e-3,1.30512e-2,-5.39764e-2,-9.67723e-2},
+        {5.84024e-2,1.81266e0,-2.59193e-1,-1.25002e1,1.99139e1,5.86545e1},
+        {2.00567e0,-7.93885e-1,-6.72461e0,3.94244e1,-1.26714e1,-1.60273e2},
+        {-1.96873e0,-5.64782e-1,1.21753e1,-3.3588e1,-4.57487e1,1.02578e2},
+        {3.5205e-1,-5.40282e-1,-5.74837e0,8.37487e0,3.28364e1,-1.06341e1}
+        }},
+        //Dy..................................................
+        CoeffMat{{
+        {1.00002e0,-9.39311e-4,5.43607e-3,1.34763e-2,-5.17249e-2,-9.56765e-2},
+        {2.78005e-2,1.79795e0,3.70737e-1,-1.24114e1,1.55759e1,5.43184e1},
+        {2.1107e0,-5.62473e-1,-9.36239e0,3.78424e1,7.94516e0,-1.36806e2},
+        {-2.02617e0,-7.62769e-1,1.52768e1,-3.08728e1,-7.18354e1,7.05554e1},
+        {3.5614e-1,-5.25061e-1,-6.86679e0,7.21247e0,4.25672e1,1.92463e0}
+        }},
+        //Ho...................................................
+        CoeffMat{{
+        {1.00004e0,-9.76747e-4,5.02266e-3,1.3853e-2,-4.89737e-2,-9.3934e-2},
+        {-3.54744e-3,1.77664e0,1.04132e0,-1.22473e1,1.08741e1,4.94269e1},
+        {2.21429e0,-3.09743e-1,-1.21254e1,3.59359e1,2.986e1,-1.11177e2},
+        {-2.07893e0,-9.74603e-1,1.84979e1,-2.77795e1,-9.928e1,3.61173e1},
+        {3.5809e-1,-5.08796e-1,-8.02233e0,5.91833e0,5.27374e1,1.53069e1}
+        }},
+        //Er...................................................
+        CoeffMat{{
+        {1.00007e0,-1.01117e-3,4.54349e-3,1.4172e-2,-4.56907e-2,-9.14834e-2},
+        {-3.54805e-2,1.74848e0,1.74921e0,-1.20063e1,5.83106e0,4.40029e1},
+        {2.3157e0,-3.56472e-2,-1.49988e1,3.37013e1,5.29727e1,-8.34879e1},
+        {-2.12613e0,-1.19953e0,2.18208e1,-2.43067e1,-1.27958e2,-6.05797e-1},
+        {3.57593e-1,-4.91966e-1,-9.2087e0,4.49274e0,6.33009e1,2.94631e1}
+        }},
+        //Tm...................................................
+        CoeffMat{{
+        {1.00009e0,-1.04219e-3,3.99979e-3,1.44301e-2,-4.18849e-2,-8.83308e-2},
+        {-6.78832e-2,1.71323e0,2.49363e0,-1.16834e1,4.45978e-1,3.80313e1},
+        {2.41433e0,2.59886e-1,-1.79765e1,3.11249e1,7.72617e1,-5.37152e1},
+        {-2.16704e0,-1.4368e0,2.52369e1,-2.04429e1,-1.57828e2,-3.9614e1},
+        {3.544e-1,-4.75067e-1,-1.04225e1,2.93269e0,7.42385e1,4.43866e1}
+        }},
+        //Yb...................................................
+        CoeffMat{{
+        {1.00012e0,-1.06927e-3,3.39091e-3,1.46208e-2,-3.75471e-2,-8.44506e-2},
+        {-1.00611e-1,1.67068e0,3.27246e0,-1.12756e1,-5.27058e0,3.15152e1},
+        {2.50948e0,5.76648e-1,-2.10477e1,2.81988e1,1.02665e2,-2.18989e1},
+        {-2.20084e0,-1.68528e0,2.87325e1,-1.61828e1,-1.88802e2,-8.08394e1},
+        {3.48226e-1,-4.58722e-1,-1.16589e1,1.23735e0,8.55166e1,6.00477e1}
+        }},
+        //Lu...................................................
+        CoeffMat{{
+        {1.00015e0,-1.09216e-3,2.71745e-3,1.47404e-2,-3.26779e-2,-7.98343e-2},
+        {-1.33495e-1,1.62059e0,4.08307e0,-1.07793e1,-1.13042e1,2.44607e1},
+        {2.60038e0,9.14519e-1,-2.42e1,2.49142e1,1.29108e2,1.19101e1},
+        {-2.22661e0,-1.94391e0,3.22924e1,-1.15206e1,-2.20784e2,-1.24202e2},
+        {3.38764e-1,-4.43546e-1,-1.29122e1,-5.9428e-1,9.70972e1,7.64128e1}
+        }},
+        //Hf..................................................
+        CoeffMat{{
+        {1.00018e0,-1.11005e-3,1.98019e-3,1.47814e-2,-2.72814e-2,-7.44728e-2},
+        {-1.66392e-1,1.56288e0,4.92359e0,-1.01915e1,-1.76472e1,1.68656e1},
+        {2.68631e0,1.27279e0,-2.74231e1,2.12642e1,1.56537e2,4.76851e1},
+        {-2.24352e0,-2.21096e0,3.59035e1,-6.45262e0,-2.53694e2,-1.69646e2},
+        {3.25734e-1,-4.30391e-1,-1.41779e1,-2.56236e0,1.08949e2,9.34555e1}
+        }},
+        //Ta..................................................
+        CoeffMat{{
+        {1.00021e0,-1.1227e-3,1.1875e-3,1.47467e-2,-2.14258e-2,-6.84575e-2},
+        {-1.99145e-1,1.49739e0,5.79226e0,-9.50812e0,-2.42938e1,8.72312e0},
+        {2.76656e0,1.65088e0,-3.07077e1,1.72391e1,1.84905e2,8.54178e1},
+        {-2.25073e0,-2.48478e0,3.95541e1,-9.72452e-1,-2.87462e2,-2.17133e2},
+        {3.0885e-1,-4.20088e-1,-1.54515e1,-4.66797e0,1.21043e2,1.11156e2}
+        }},
+        //W...................................................
+        CoeffMat{{
+        {1.00025e0,-1.12974e-3,3.25083e-4,1.46196e-2,-1.49818e-2,-6.15916e-2},
+        {-2.31499e-1,1.42403e0,6.68301e0,-8.72894e0,-3.12005e1,8.13039e-2},
+        {2.84004e0,2.04799e0,-3.40292e1,1.28426e1,2.14035e2,1.24907e2},
+        {-2.24701e0,-2.76343e0,4.3216e1,4.91177e0,-3.2188e2,-2.66423e2},
+        {2.87708e-1,-4.13579e-1,-1.67235e1,-6.90724e0,1.33305e2,1.29426e2}
+        }},
+        //Re..................................................
+        CoeffMat{{
+        {1.00028e0,-1.13018e-3,-5.85208e-4,1.44008e-2,-8.12721e-3,-5.40793e-2},
+        {-2.63328e-1,1.34285e0,7.59541e0,-7.85118e0,-3.83723e1,-9.07534e0},
+        {2.90612e0,2.46267e0,-3.73829e1,8.06978e0,2.43918e2,1.66172e2},
+        {-2.23164e0,-3.04426e0,4.68824e1,1.12007e1,-3.56918e2,-3.17506e2},
+        {2.62068e-1,-4.12056e-1,-1.79911e1,-9.27922e0,1.45719e2,1.48255e2}
+        }},
+        //Os..................................................
+        CoeffMat{{
+        {1.00032e0,-1.12442e-3,-1.55891e-3,1.40881e-2,-7.3226e-4,-4.57754e-2},
+        {-2.94323e-1,1.25374e0,8.52167e0,-6.87354e0,-4.57537e1,-1.8691e1},
+        {2.96357e0,2.89403e0,-4.07385e1,2.92128e0,2.74335e2,2.08986e2},
+        {-2.20324e0,-3.32515e0,5.05197e1,1.78893e1,-3.92323e2,-3.70113e2},
+        {2.31474e-1,-4.16536e-1,-1.92429e1,-1.1781e1,1.58197e2,1.67545e2}
+        }},
+        //Ir...................................................
+        CoeffMat{{
+        {1.00036e0,-1.11107e-3,-2.58076e-3,1.36692e-2,7.08911e-3,-3.6773e-2},
+        {-3.24316e-1,1.15695e0,9.45947e0,-5.79675e0,-5.33328e1,-2.87528e1},
+        {3.01159e0,3.33981e0,-4.4085e1,-2.59468e0,3.05217e2,2.53266e2},
+        {-2.16093e0,-3.60244e0,5.41141e1,2.49638e1,-4.28002e2,-4.24126e2},
+        {1.95645e-1,-4.28567e-1,-2.04742e1,-1.44066e1,1.70702e2,1.87249e2}
+        }},
+        //Pt...................................................
+        CoeffMat{{
+        {1.00039e0,-1.09011e-3,-3.65347e-3,1.31413e-2,1.53649e-2,-2.70335e-2},
+        {-3.53028e-1,1.05257e0,1.04028e1,-4.62034e0,-6.10681e1,-3.92186e1},
+        {3.04907e0,3.79831e0,-4.73979e1,-8.47391e0,3.36394e2,2.98829e2},
+        {-2.10347e0,-3.87306e0,5.76385e1,3.24147e1,-4.63751e2,-4.79318e2},
+        {1.54174e-1,-4.49506e-1,-2.16756e1,-1.71516e1,1.83161e2,2.07283e2}
+        }},
+        //Au...................................................
+        CoeffMat{{
+        {1.00043e0,-1.06114e-3,-4.7734e-3,1.25001e-2,2.40713e-2,-1.65728e-2},
+        {-3.80207e-1,9.40778e-1,1.13464e1,-3.34423e0,-6.89259e1,-5.00549e1},
+        {3.07493e0,4.26732e0,-5.06567e1,-1.47102e1,3.67721e2,3.45521e2},
+        {-2.0297e0,-4.13334e0,6.10696e1,4.02298e1,-4.99399e2,-5.35495e2},
+        {1.06694e-1,-4.80932e-1,-2.28393e1,-2.00106e1,1.95512e2,2.27574e2}
+        }},
+        //Hg.................................................
+        CoeffMat{{
+        {1.00047e0,-1.02377e-3,-5.93877e-3,1.17393e-2,3.32021e-2,-5.38025e-3},
+        {-4.05577e-1,8.21834e-1,1.22846e1,-1.96909e0,-7.68675e1,-6.12207e1},
+        {3.08808e0,4.74434e0,-5.38383e1,-2.12945e1,3.99035e2,3.93158e2},
+        {-1.93842e0,-4.37923e0,6.43816e1,4.83938e1,-5.3475e2,-5.92428e2},
+        {5.28167e-2,-5.24544e-1,-2.39565e1,-2.29768e1,2.07685e2,2.48037e2}
+        }},
+        //Tl.................................................
+        CoeffMat{{
+        {1.00051e0,-9.77691e-4,-7.14416e-3,1.08563e-2,4.27196e-2,6.50699e-3},
+        {-4.28882e-1,6.96049e-1,1.32126e1,-4.94649e-1,-8.48638e1,-7.26907e1},
+        {3.08745e0,5.22659e0,-5.69235e1,-2.82201e1,4.30208e2,4.41611e2},
+        {-1.82849e0,-4.60634e0,6.7553e1,5.68935e1,-5.69648e2,-6.49946e2},
+        {-7.81506e-3,-5.82177e-1,-2.50201e1,-2.60444e1,2.19621e2,2.68607e2}
+        }},
+        //Pb.................................................
+        CoeffMat{{
+        {1.00055e0,-9.22411e-4,-8.38967e-3,9.83942e-3,5.26394e-2,1.91419e-2},
+        {-4.49833e-1,5.63865e-1,1.41246e1,1.0768e0,-9.28737e1,-8.44177e1},
+        {3.07189e0,5.71077e0,-5.98888e1,-3.54715e1,4.61066e2,4.9067e2},
+        {-1.69869e0,-4.80968e0,7.05579e1,6.57062e1,-6.03887e2,-7.07793e2},
+        {-7.55851e-2,-6.55871e-1,-2.60213e1,-2.92042e1,2.31249e2,2.89188e2}
+        }},
+        //Bi.................................................
+        CoeffMat{{
+        {1.00059e0,-8.58152e-4,-9.6602e-3,8.70229e-3,6.28328e-2,3.23376e-2},
+        {-4.68192e-1,4.25668e-1,1.50174e1,2.74763e0,-1.00883e2,-9.64012e1},
+        {3.04043e0,6.19366e0,-6.27207e1,-4.30487e1,4.91532e2,5.40291e2},
+        {-1.54799e0,-4.98431e0,7.3381e1,7.48252e1,-6.37365e2,-7.65885e2},
+        {-1.5081e-1,-7.47659e-1,-2.69551e1,-3.24524e1,2.42528e2,3.09744e2}
+        }},
+        //Po.................................................
+        CoeffMat{{
+        {1.00062e0,-7.84582e-4,-1.09553e-2,7.43533e-3,7.33106e-2,4.61358e-2},
+        {-4.83646e-1,2.8205e-1,1.58842e1,4.51493e0,-1.08846e2,-1.08589e2},
+        {2.99185e0,6.67128e0,-6.53938e1,-5.09329e1,5.21419e2,5.90246e2},
+        {-1.37512e0,-5.12439e0,7.59949e1,8.42238e1,-6.69862e2,-8.23945e2},
+        {-2.33878e-1,-8.59879e-1,-2.78125e1,-3.57783e1,2.53381e2,3.30173e2}
+        }},
+        //At.................................................
+        CoeffMat{{
+        {1.00066e0,-7.01484e-4,-1.22684e-2,6.03454e-3,8.40298e-2,6.05018e-2},
+        {-4.95899e-1,1.3369e-1,1.67195e1,6.37574e0,-1.16723e2,-1.20936e2},
+        {2.92499e0,7.13921e0,-6.78861e1,-5.91051e1,5.50562e2,6.40337e2},
+        {-1.17889e0,-5.22361e0,7.8376e1,9.38748e1,-7.01185e2,-8.81728e2},
+        {-3.25146e-1,-9.95032e-1,-2.85858e1,-3.91713e1,2.63738e2,3.50385e2}
+        }},
+        //Rn.................................................
+        CoeffMat{{
+        {1.0007e0,-6.08729e-4,-1.35873e-2,4.50061e-3,9.48991e-2,7.53277e-2},
+        {-5.04647e-1,-1.86445e-2,1.75175e1,8.32729e0,-1.24477e2,-1.33405e2},
+        {2.83871e0,7.59269e0,-7.01766e1,-6.75463e1,5.78809e2,6.90388e2},
+        {-9.58143e-1,-5.27523e0,8.0502e1,1.03751e2,-7.31155e2,-9.39013e2},
+        {-4.24964e-1,-1.15578e0,-2.92679e1,-4.26205e1,2.73537e2,3.70298e2}
+        }},
+        //Fr.................................................
+        CoeffMat{{
+        {1.00073e0,-5.06594e-4,-1.49195e-2,2.82752e-3,1.05984e-1,9.07179e-2},
+        {-5.09498e-1,-1.74114e-1,1.82691e1,1.0364e1,-1.32042e2,-1.45912e2},
+        {2.73154e0,8.02658e0,-7.22315e1,-7.62278e1,6.05902e2,7.40063e2},
+        {-7.11399e-1,-5.27209e0,8.23375e1,1.13815e2,-7.59479e2,-9.95409e2},
+        {-5.33773e-1,-1.34491e0,-2.98474e1,-4.61115e1,2.82675e2,3.89772e2}
+        }},
+        //Ra.................................................
+        CoeffMat{{
+        {1.00076e0,-3.94938e-4,-1.62461e-2,1.01616e-3,1.17156e-1,1.06529e-1},
+        {-5.10179e-1,-3.3169e-1,1.89693e1,1.24811e1,-1.39382e2,-1.58412e2},
+        {2.60247e0,8.43503e0,-7.4033e1,-8.51228e1,6.31692e2,7.89166e2},
+        {-4.37673e-1,-5.2062e0,8.38642e1,1.2403e2,-7.85984e2,-1.05067e3},
+        {-6.5185e-1,-1.56551e0,-3.03186e1,-4.96307e1,2.91091e2,4.08719e2}
+        }},
+        //Ac.................................................
+        CoeffMat{{
+        {1.00079e0,-2.74014e-4,-1.75614e-2,-9.33046e-4,1.28369e-1,1.22714e-1},
+        {-5.06317e-1,-4.90282e-1,1.96102e1,1.46725e1,-1.4644e2,-1.70838e2},
+        {2.45014e0,8.81187e0,-7.5553e1,-9.42005e1,6.55963e2,8.37419e2},
+        {-1.35635e-1,-5.06921e0,8.50532e1,1.34356e2,-8.10425e2,-1.10448e3},
+        {-7.79579e-1,-1.82077e0,-3.06725e1,-5.31628e1,2.987e2,4.2702e2}
+        }},
+        //Th................................................
+        CoeffMat{{
+        {1.00082e0,-1.43895e-4,-1.8859e-2,-3.02404e-3,1.39581e-1,1.39236e-1},
+        {-4.97537e-1,-6.48585e-1,2.01838e1,1.69296e1,-1.53157e2,-1.83108e2},
+        {2.27321e0,9.15011e0,-7.6763e1,-1.0342e2,6.78484e2,8.84496e2},
+        {1.9599e-1,-4.85189e0,8.5876e1,1.44742e2,-8.32543e2,-1.15645e3},
+        {-9.17319e-1,-2.1142e0,-3.09003e1,-5.66889e1,3.05413e2,4.44542e2}
+        }},
+        //Pa.................................................
+        CoeffMat{{
+        {1.00085e0,-5.30411e-6,-2.01239e-2,-5.2478e-3,1.50683e-1,1.55963e-1},
+        {-4.83526e-1,-8.05262e-1,2.06846e1,1.92461e1,-1.59496e2,-1.95177e2},
+        {2.07059e0,9.44256e0,-7.76442e1,-1.12749e2,6.99103e2,9.30193e2},
+        {5.58224e-1,-4.54471e0,8.63141e1,1.55146e2,-8.52163e2,-1.20633e3},
+        {-1.06534e0,-2.44938e0,-3.09967e1,-6.01932e1,3.11168e2,4.61192e2}
+        }},
+        //U....................................................
+        CoeffMat{{
+        {1.00087e0,1.42049e-4,-2.13455e-2,-7.61031e-3,1.61591e-1,1.72814e-1},
+        {-4.6392e-1,-9.58725e-1,2.11056e1,2.16134e1,-1.65402e2,-2.06974e2},
+        {1.841e0,9.68106e0,-7.81727e1,-1.22145e2,7.17614e2,9.74218e2},
+        {9.52222e-1,-4.13715e0,8.63446e1,1.65515e2,-8.69054e2,-1.25378e3},
+        {-1.22395e0,-2.83024e0,-3.09548e1,-6.36564e1,3.15887e2,4.76849e2}
+        }}};
+    // clang-format on
+    static_assert(std::size(mott_coeffs) == MottElementData::num_mott_elements,
+                  "wrong number of Mott coefficient elements");
+
+    int index = z.unchecked_get() - 1;
+    CELER_VALIDATE(
+        index >= 0 && index < int{MottElementData::num_mott_elements},
+        << "atomic number " << z.get()
+        << " is out of range for Coulomb scattering model Mott coefficients "
+           "(must be less than "
+        << MottElementData::num_mott_elements << ")");
+
+    return mott_coeffs[index];
+}
+
+//---------------------------------------------------------------------------//
+}  // namespace celeritas
