+++ src/celeritas/em/process/CoulombScatteringProcess.cc
@@ -19,13 +19,10 @@ namespace celeritas
 /*!
  * Construct from host data.
  */
-CoulombScatteringProcess::CoulombScatteringProcess(
-    SPConstParticles particles,
-    SPConstMaterials materials,
-    SPConstImported process_data,
-    CoulombScatteringModel::Options const& options)
+CoulombScatteringProcess::CoulombScatteringProcess(SPConstParticles particles,
+                                                   SPConstImported process_data,
+                                                   Options const& options)
     : particles_(std::move(particles))
-    , materials_(std::move(materials))
     , imported_(process_data,
                 particles_,
                 ImportProcessClass::coulomb_scat,
@@ -33,8 +30,6 @@ CoulombScatteringProcess::CoulombScatteringProcess(
     , options_(options)
 {
     CELER_EXPECT(particles_);
-    CELER_EXPECT(materials_);
-    CELER_EXPECT(options_);
 }
 
 //---------------------------------------------------------------------------//
@@ -45,7 +40,7 @@ auto CoulombScatteringProcess::build_models(ActionIdIter start_id) const
     -> VecModel
 {
     return {std::make_shared<CoulombScatteringModel>(
-        *start_id++, *particles_, *materials_, options_, imported_.processes())};
+        *start_id++, *particles_, imported_.processes())};
 }
 
 //---------------------------------------------------------------------------//
