+++ src/celeritas/phys/ProcessBuilder.hh
@@ -110,7 +110,6 @@ class ProcessBuilder
   private:
     //// DATA ////
 
-    ImportData const& data_;
     UserBuildInput input_;
     UserBuildMap user_build_map_;
     std::function<ImportSBTable(AtomicNumber)> read_sb_;
@@ -119,6 +118,8 @@ class ProcessBuilder
 
     BremsModelSelection selection_;
     bool brem_combined_;
+    bool enable_lpm_;
+    bool use_integral_xs_;
 
     //// HELPER FUNCTIONS ////
